backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
sqlalchemy>=2.0,<3
asyncpg>=0.29,<1
pydantic-settings>=2.4,<3
claude-code-sdk
